requirements.txt
astroid==1.6.3
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
Django==2.0.4
django-pure-pagination==0.3.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.8.4
pytz==2018.4
pywebber==5.0
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
